Flatten log elements; split up raw logs in HTML
I was logging the list of properties of an object as a property called
'props'. I used to use just the object properties in the HTML log
viewing page. Now I need access to a bunch of extra log properties so
I'm just flattening out the object properties and making them top level
log properties. I may not need the flatten/1 function: I noticed that I
was adding {props, Props} to the list, which was making things screwy,
so I've just concatenated on the object props instead.

To make this easier to figure out when I'm printing out the raw log
lines in HTML I put each key+value pair in its own div.