Add filters to the JSON log HTML page
When debugging I need to know what messages are being passed around and
handled; however, many starting messages are "link" and "populate" which
I'm not usually interested in. Other messages are not handled by the
object that received them and, while occassionally useful--i.e. when an
object that should have handled it ignored it instead--they are often
just noise.

Each filter, when activated, clears all the elements from the page,
filters the list of JSON log entries according to all the activated
filters, then recreates all the appropriate page elements.

Any applied filters are re-applied once the page is re-populated.

I could probably preserve all the filter elements on the page and only
recreate the JSON log elements.

In the previous incarnation of HTML logging (instead of logging to JSON
and turning the JSON into HTML) I was building each HTML element so that
it could be turned off in any one of several layers. Any layer that was
turned off (display set to 'none') would hide any inner layers as
well as the innermost content. Confusing the content with the structure
like that was very inflexible. I knew better but I wanted to get
something out quickly and it proved to be immensely valuable for a
while. This should be flexible in that both the same client and
different clients can display the raw content however they like. (This
is the entire point of having HTML and CSS be separate).