Log mouseover span, show message, show non-pids
Instead of showing the log text div when you mouseover the entire log
div there's now a dedicated "log" span. Mouseover the span and the
entire raw log div shows up. I've also changed the background of the log
text div to white so that it doesn't blend with the log div text.

Add the text version of the message at the end of each log div. The
event span only shows "<source> <event> <target>" but some messages
don't fit that format, such as resource reservation events.

Show non-PID atoms in the "source event target" event span when either
the source or the target hasn't yet been replaced with a pid. For
example, when a player types "attack zombie", 'zombie', the atom will
get replaced with the pid of a zombie process in the room, if there is
one. When we log the message from the player, if the atom 'zombie' isn't
display then all we see in the log web page is "Player ZZ attack __",
which isn't super helpful.