Logging changes: smaller font, fix handler, etc.
Use smaller fonts for all parts of the log message to take up less
space.

Use the 'handler' key to get the message handler instead of 'module',
which was always erlmud_object.

Set some sizes so that things line up properly on the log line even if
those log property don't have values.

Fill in the room property during attempt processing. This isn't working
right and needs some help.

When a 'resend' is done add the new message after the original message.
This reminded me that sometimes handlers don't resend a message but
rather kick off a completely new message: this would be necessary if you
want multiple attack vectors to each see an attack message. If the
message {Pid1, attack, Pid2} was resent by the first attack vector to
{Pid1, attack, Pid2, with, AttackVector1} then no other attack vector
would have a chance to respond.