Fix plumbing for log properties
The fail handling wasn't accounting for handlers to be able to return
either {Props, LogProps} or just Props.

While working on this I had to review why, when an object receives a
notification that an event failed, does it call every handler with that
failed message. If only one handler can subscribe to that event,
shouldn't we store which handler subscribed and only notify that
handler? While I think that might be a viable option, the way I
currently handle it is that event handlers listen for particular
messages whether during the attempt, success or failure. It's possible
to have a handler that didn't original subscribe to the event respond on
failure. I'm not sure if that will be useful or not.