Allow changed ID, ID must be atom, bin fail reason
Fix remaining tests. All tests pass again!

Some tests start new objects with no ID so one needs to be generated; if
it's generated then it's going to be different than the 'undefined' that
was passed in--hopefully--so the same variable can't be used to bind the
new value. If the ID is passed in then it _is_ the same value, which is
why I didn't catch this at first.

Object IDs must be atoms according to the erlmud_index guard clauses. I
have no idea what this is, except that erlmud_event_log expects them to
be atoms.

Allow for binary fail reasons, especially since we convert string and
atom fail reasons to binaries anyway.