Check for mising log file, fix purging
The log web page is now being used for live logging and for reading log
files. If there is no log file I don't want an error so I'm putting log
file logs in a separate variable and checking for its existence before
trying to use it.

For the log purging, if we get more than 10 logs but then try and delete
everything after the 99th log then obviously we're going to get an
error. Until we actually get logs 11 through 100 we're not going to have
anything to delete. Just a dumb typo.