Dump the world to console using PID from index
It looks like at some point I was registering each object process with
its ID. Since I'm not doing that anymore I'm using the erlmud_index to
look up the pid from the ID in the play world. With the pid I can look
up the objects gen_server state.

I also did a bit of work to make the move less buggy but I didn't
actually try it. I'd like to be able to run all of the commands through
the websocket connection.