Centralize log flattening in erlmud_event_log
I need to flatten logs that aren't initiated in erlmud_object so I've
moved the code to erlmud_event_log.

A lot of details get added automatically for attempts, successes and
failures. For ad hoc logging the properties have to be added by hand.
The properties also need to be flattened manually.