Fix up look command
Integer values for attributes were getting garbled.

Old code from a previous method of handling descriptions was left around
in the play world: the "desc" property no longer has keywords to
describe which values should be filled in in the description;
description templates are now stored in the app.src file.