WIP: spells, spell effects, testing
Spells have spell effect "prototypes" that spawn spell effects. The
spells can auto-attack or be triggered manually. Once triggered the MUD
will calculate first whether the spell can be cast; if the spell is cast
then the prototype spell effect object creates a copy of itself and
attaches it to the target. The spell effect can be instantaneous or it
can be lingering. Eventually the spells can be defensive or offensive.
Spells must be memorized before they can be cast. Once the spell effect
is created then it will calculate if it can hit and then if it can do
damage.

Also working on testing the spells.

There is some difficulty sending results to the player when the spell
effect doesn't know the name of the target. I'm thinking the players
connection object can take messages with {PID, property} tuples and
gather the information to fill them in.

e.g. [<<"You hit ">>, {TargetPid, name}, <<" with fireball ...">>]
The connection object can send a message to TargetPid to get its name
property.

Changing hit and damage calculations to use a "base" value (e.g. attack
hit, attack damage) to calculate a random "roll" then applying modifiers
on top of that.