Add icons, allocate source, clean up CSS
Added the meter, spell and effect icons to make the logs easier to
parse. Also added the resource PID to the allocation message so that
allocation messages could be logged with both a source and a target,
giving each icons.

Split the handler module names and took everything after
"erlmud_handler_" since it's redundant. Changed the width of the process
boxes and the "event" box so that processes wouldn't wrap.