WIP: back up gerlshmud in Mnesia
I want mud objects to restart if they crash. The process will
continually store it's properties in Mnesia and when it crashes it will
restart and look for it's properties in Mnesia. Properties that point to
processes will be stored with the property key and the ID of the
process. The process PID will be looked up in Mnesia as well.

I'll need to move the PID storage from erlmud_index to Mnesia. However,
I might just make erlmud_index a front-end to Mnesia.