Don't start mnesia right away, set dir, debugging
Mnesia can't be started by setting it as one of the required
applications gerlshmud.app.src because we have to create the schema
_before_ Mnesia starts: if we don't it gives a weird error about the
schema already existing.

Not that when the tests run, since the mnesia directory is relative, the
directory will be in logs/ct_run.ct_gerlshmud\@...

Also note that in order to read what's in the test Mnesia tables the
node name has to be the same: 'ct_gerlshmud'.

This is all to support restarted crashed processes using their saved
properties in Mnesia. The PIDs will have to be looked up using
gerlshmud_index (which should be in Mnesia as well).