WIP: replacing list index with Mnesia object table
If I'm going to store object properties in Mnesia, which is backed by an
in-memory ETS table then I might as well use it to look up PIDs and IDs
as well rather than use a process with a raw list.

This commit is a lot of trimming, organizing, sorting and refactoring to
mix the old index code with Mnesia calls and dead PID logic.

When a process dies all of the objects that have links to that process
will be notified since they all trap exits. Each linked object will
register to be notified of the new PID when a new process is started and
loaded back up with the original properties (that are stored in Mnesia).

Since there's a race condition with the new process being started to
re-host the orphaned properties and linked objects registering to be
notified there two tables: one to store subscriptsions and one to store
the mapping of old PID to new PID. If the process starts before an
object can subscribe for the new PID then the mapping record is used to
inform the linked object subscriber right away. If the linked object
subscriber gets a subscription record in first then once the replacement
process is started then we'll check for subscriptions. In case they
don't get deleted each record gets a send_after call to delete them
after a minute, which should be plenty of time to get things sorted.

TODO: once the linked object subscriber is notified of the new PID it
needs to link to it.

There's a flaw in this: if a process is dying because of bad properties
then resurrecting it with a new process will probably just cause it to
crash again, eventually bringing down the supervisor and thus the whole
server.

When a process dies all linked objects will mark properties with the PID
of that process as {dead, PID}. These PIDs are not converted to IDs when
being stored in Mnesia so that they can be compared to the new PID when
the dead object is restarted in a new process.

Since a restarted process needs to notify linked objects of it's new
PID, and since it needs to send it's old PID so the linked objects will
know what to replace, the object stores it's current PID in it's
properties. However, the index shouldn't convert this PID to an ID as
that would defeat the purpose of keeping the old PID around.