Convert IDs to PIDs once and link at the same time
On object start if the object didn't have properties in Mnesia then the
properties were stored in Mnesia and then all IDs were converted to PIDs
using whatever was in Mnesia. This is what the populate function was
doing that's called by the gerlshmud_object_sup supervisor, except
populate _also_ links the process. If the IDs are converted to PIDs in
the start function but not linked then populate doesn't link them and
then that object misses the message that a linked object has died. This
fixes the revive process test.