Many effects are not damage
Converting _effect_attack over to use a generic "roll" value for the
amount of effect that was produced: this could be negative, zero or
positive, depending on other objects that affected the effect (e.g. by
blocking or mitigating it).

I'm removing the spell attack handler because it can now be handled by a
generic attack handler and simply specify the attack type as "spell".

I don't plan to output anything to the player if an effect is not
applied to a target. This would just be noise. If you don't see a
message than nothing happened. When something _does_ happen, it will
stand out.

I'm changing the output message to include the character and the target
as well as embedding PIDS into the message that will need to be filled
in the PID processes. The rooms should also send the message. The GUI or
colour coding could perhaps display the attack in adjacent rooms with a
more muted colour, smaller font, etc.