Attacks will need types for defence objects to check
In order for a defence object, e.g. a shield to know if it should
deflect an attack it needs to know what type of attack it was.

I'm planning to limit each attack to one type for hitting, but then have
multiple effects for dealing out the damage. e.g. a burning club will be
"melee" for hitting, but then maybe "bash" and "fire" for damage
effects. Each defence object will need to know which attack types and
which effects it can defend against.