Attacks roll hits, effects cause damage
An "attack" is anything that can be used to attack another entity in the
game, e.g. a fist, a spell, gun, laser, etc. We calculate whether an
attck succeeded, for example, whether a sword swing connected with the
target. If an attack succeeds then a process is started for each one of
the attacks "effects".

An "effect" is a change, typically damage, caused by a successful attack.
Each effect linked to an attack gets its own process and proceeds to
calculate how much of the effect is applied to the target, if any. For a
sword swing the effect might be "cut"; for a spell the effect might be
"fire", and "blunt force".