WIP: store effect properties in attack process
Since attack processes create the effect processes that will carry out
the attack the effect properties will be stored with the attack process.

I thought about having an effect prototype process that stored the
properties and having it replicate itself on demand, but then I'd have
two different effect process which would need to respond to some effect
messages, but not others. This might not actually be all that confusing
at all, but, whatever, I'm going this way for now.

The attack process needs to allow for multiple effects so it will need
to store an list of proplists instead of having single properties for
each aspect of a single effect.

Actually, different effects should be different processes. I'm thinking
of a flaming sword: that should be a regular sword with a burning effect
attached to it. I'll need to add child attacks that respond to the
parent. I think I'm already modelling this with bullets: a bullet is a
child process of an attacking parent process.