Create effects with 'child_handlers' as handlers
When a prototype effect creates an effect instance it uses its own
properties (i.e. the prototype for an effect) _EXCEPT_ the handlers: the
prototype listens for events to create effects, the 'child_handlers'
property will listen for events to apply effects.

So far the child_handlers property will be a tuple of
{handlers, ListOfEffectApplicationHandlers}
Which can just be appended to the effect instance properties _instead_
of the effect creator process handlers property.