WIP: moving most logic from attack to effect
Moving the attacks over to various effects still makes sense; however,
moving resource allocation doesn't: are there any effects of an attack
that have different resources? If so, how do we coordinate them? If the
"slice" effect of a sword has the resources it needs, does it wait for
whatever resources "shock", "burn" or "poison" might need? (A creative
person could always come up with a scenario where something like this
could be possible). Having a parent effect wait for the resources
required for a dpeendent effect sounds like over-engineering.

My go-to example is a flaming sword: it has a main effect of slicing (or
whatever) but it has a secondary, dependent effect of burning (if you
don't hit with the sword then you also don't burn with it.

Does the "burn" effect have it's own resources? Does it need to request
resources every time it applies itself (if it's an on-going/repeating
effect)?

I don't think so. The idea of an "attack" is that it's some kind of
offensive action like a weapon, or a spell. We can't activate the
offensive without the appropriate resources required to kick off _any_
of the effects. So, the attack should wait for resources (e.g. stamina
and focus for a sword) and then tell the effects to kick off. Dependent
effects will wait for parent effects to hit first before kicking into
action.

So, I'm going to move the resources back to the attack handler from the
effect handler.

Having effects handle their own rolls to hit and calculate effect
amount, and having the hits cascade (i.e. the dependency: if A hits then
B might hit) still makes sense.

NOTE:

I need to remember that I'm ignoring time, CPU and transactionality:
things don't have to happen exactly in order and groups of effects don't
need to be atomic: if you manage to shock something twice when you've
only hit it with a sword once, well good for you. Life is messy.