Fix bugs, don't "affect" in effect handler
The effect handler rolls for a hit and then if it hits it rolls for
effect amount (e.g. damage amount): it does not have anything it can do
with this knowledge, only the processes belonging to the target can do
something with this: e.g. the target's hitpoints process could knock off
(or add, depending on the effect) some hitpoints.

Just noodling: a weapon could steal health, in which case the attacker
would need to listen for effects as well.