Wait for conditions functions
Create a wait loop while we wait for conditions to be met which are
encoded as functions with a description. I don't know how long we'll
need to wait for things that require resources so we'll just wait for a
reasonable amount of time, but we'll succeed as soon as possible.