Attacks create effects, effects do damage
Finish the logic for having attacks sending off "affect" messages when
they have enough resources.

The effect prototype process creates new effect processes using a copy
of it's own properties with the target added and the handlers replaced
with 'child_handlers'. This is so the effect process doesn't act like an
effect prototype process.

The effects do hit and amount rolls. Once the hit roll succeeds and the
amount roll is positive then a message is sent out to apply the effect.

For now I'm just having the hp process listen for 'blunt_force' effects,
but I'll probably need to make this configurable: probably a list of
effect types that cause HP to be lost.

Currently effects can't be on-going (like burning) since they don't
re-run themselves.

Once an effect process is finished it will instruct other processes to
delete itself so that it's no longer part of the graph. The delete is
done through a new universal handler that activates on attempt, not
success: there's no need to wait for any processes to vet the attempt:
nothing can cancel the deletion of a property as it's part of the graph
logic, not the game logic. Cancelling an attempt is something that game
elements can do to signal that something hasn't worked: e.g. opening a
door that's locked, trying to grab a non-existent object, etc.