Fix player_resource_wait test
Damage is now done by effects and the amount of damage is random. This
updates the test to make sure _some_ damage has been done, not a
specific amount.

This could fail if the effect can do more damage than the enemy has
because I only set a long wait time on the necessary resource: i.e. we
can hit once; it that hit is _really_ damaging the target could die.