Attach necessary module properties
In order for an object to participate in attack and defence it needs to
have the appropriate 'should_attack_module' and 'should_defend_module'.
These control whether this attack or defence is ready to use. Melee
weapons need to be wielded and spells need to be memorized.

In order for an attack to do anything it has to have an
'effect_prototype' property. This is the object that will spawn the
effect processes. If there's no effect prototype object listening for
attacks then no effects will be spawned.