Add protocol extraction parse_transform to all handlers
There's a parse_transform that finds all attempt and succeed calls, as
well as calls to gerlshmud_object:attempt/2. It extracts the events out
of the calls and writes them to a CSV file so I can see what protocol
has emerged and to see any deviations between modules. Often it's very
hard to debug why a message created by one handler doesn't trigger
another handler; often it's because the messages are slightly different.

With this CSV I can sort by the events to see which modules have similar
events and compare them.