Find resends and broadcasts
I wanted to see where events are modified and resent, as well as where
they are broadcast. Since most resends and broadcasts build up a
"NewMessage" variable first and then return the tuple, I need to first
find "NewMessage" and then when I see the resend/broadcast, record that
with the contents of "NewMessage". In order to do this I've ditched the
old technique of keeping a process for each file to track the module and
I've threaded a State map through the recursive search/2 calls. This was
a giant pain.

Broadcasts still aren't printing out in the 7th column but resends are.

I've renamed expr/2 to search/2 to reflect that I'm just searching
through the AST.

I've renamed bin/1 to print/1 to reflect that I'm printing something out
based on what I've found in the AST.

I had actually collapsed search/2
and print/1 into a single function before I realized that I only want to
recursively print in certain situations: e.g. when I've found a attempt
clause that isn't ignored.