Link attempts with broadcasts or resends
If an attempt resends a different message, or broadcasts a message, then
I want to know that one event triggered another.

Likewise, if there's a broadcast or a resend then I want it recorded on
it's own line and I want it linked back to the attempt that kicked it
off.