WIP: stop characters and move items
When a character dies set a timer for a corpse cleanup; cleanup will
trigger items to move from the character's corpse to the room. Once the
character's items are cleaned up the character's processes should all
stop.

In order to stop cleanly the supervisor's terminate function is called,
but the object gen_servers catch the exit and exit with reason normal so
that we don't get a bunch of errors in the console.

All the character process (e.g. hp, life, etc.) are dying, but I don't
know why: I'm trying to trace the handler functions but I'm not seeing
the calls I expect: e.g. _char_cleanup doesn't get {Self, cleanup}.

I added some process-specific tracing to gerlshmud_dbg so I could limit
it to particular objects.