WIP: fix promiscuous handler succeed
Update the stop handler to only return {stop, ...} when succeed is
called for the {stop, Character} event.

If a process subscribes to an event then every handler for that process
has its succeed/2 function called; gerlshmud_object doesn't track which
handler subscribed to the event. In fact, you can subscribe to an event
in one handler and handle the success in:
- another handler
- multiple handlers
- no handlers
  - Not materially; every handler has to at least return the Props arg.