Add body_part cleanup handler to body parts
Body parts need to add themselves to the cleanup even so that items held
by body parts know whether the corpse they are on is being cleaned up. I
could add a character property to the item when it's added to a child in
the character graph, and remove it when the item is removed, but that
feels more complicated.