Clarify comments
I read these comments and didn't realize that I was referring to
sub-items moving away from parent items. Otherwise it seems weird that
"Self" is the target in an inventory move handler: if it was the "top"
item then the source and target of the move wouldn't be items: they'd be
a body_part, a character, a room, etc.

To be clear: items can have sub-items: e.g. a gun might have a bullet; a
necklace might have a gem stone.