Query log lines with jq
I can view the log web page or view the JSONP file in the browser
(formatted nicely), but these are clunky.

With jq I can whip up a query and see only the fields that I want right
in the terminal.

"Make it work, make it great, make it fast."
This works. :)