Don't need new messages
Any body_part will add itself to the message as soon as it sees it: we
don't need to resend the message in case some process let it go by
without recognizing it.

When we replace a name with a PID then we'd need to resend the message
because no handlers except the "..._inject_self" handler would recognize
the message: all handlers need to see the updated message.