Don't need skin, char vs. room items
I don't need a "don't drop on death" item because searching a character
or a corpse has nothing to do with what's left on the ground after a
character's corpse is cleaned up.

I need to make sure that both items on the character and items on
character body parts are both listed in the search. I don't need three
items on the body but ... whatever.

Fix the book name: doesn't need to be "book1"