Fix guard boolean expression
From the Erlang reference manual:
"A guard sequence is a sequence of guards, separated by semicolon (;).
 The guard sequence is true if at least one of the guards is true.
 (The remaining guards, if any, are not evaluated.)"

In order to have X and (Y or Z) I need to use "orelse" for the or,
instead of semi-colons; semi-colons won't work in parentheses.

I had this in the char inject self but I guess I missed it in the item
inject self.